homework9/services/auth/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djoser
psycopg2